# gen4: Fast aggregations