… go/vt/vtgate/engine/aggregations_test.go
… go/vt/vtgate/engine/opcode/constants_test.go
… go/vt/vtgate/engine/ordered_aggregate_test.go
… go/vt/vtgate/engine/scalar_aggregation.go
… go/vt/vtgate/evalengine/api_aggregation_test.go
… go/vt/vtgate/evalengine/api_arithmetic_test.go
… go/vt/vtgate/evalengine/api_compare.go
… go/vt/vtgate/evalengine/api_hash_test.go
… go/vt/vtgate/evalengine/weights_test.go