@dbussink 2023-08-31
@vmg There's a number of test failures here that seem related?

```
=== RUN   TestBuggyQueries
    mysql.go:219: Query (select /*vt+ PLANNER=Gen4 */ min(tbl0.loc) from dept as tbl0) results mismatched.
        Vitess Results:
        [VARCHAR("CHICAGO")]
        MySQL Results:
        [VARCHAR("BOSTON")]
        query plan: 
        {
        	"OperatorType": "Aggregate",
        	"Variant": "Scalar",
        	"Aggregates": "min(0 COLLATE utf8mb4_general_ci) AS min(tbl0.loc)",
        	"Inputs": [
        		{
        			"OperatorType": "Route",
        			"Variant": "Scatter",
        			"Keyspace": {
        				"Name": "ks_random",
        				"Sharded": true
        			},
        			"FieldQuery": "select min(tbl0.loc) from dept as tbl0 where 1 != 1",
        			"Query": "select /*vt+ PLANNER=Gen4 */ min(tbl0.loc) from dept as tbl0",
        			"Table": "dept"
        		}
        	]
        }
  
```

They are all for aggregate functions so I think there's an issue here somewhere.