@harshit-gangal 2023-09-01
error looks legit

```
2023-08-31T15:48:53.0284328Z         Vitess Results:
2023-08-31T15:48:53.0284535Z         [VARCHAR("ACCOUNTING") VARCHAR("cattle")]
2023-08-31T15:48:53.0284656Z         MySQL Results:
2023-08-31T15:48:53.0284865Z         [VARCHAR("SALES") VARCHAR("cattle")]
```